+++ cassandra/query.py
@@ -72,8 +72,9 @@ def tuple_factory(colnames, rows):
     Example::
 
         >>> from cassandra.query import tuple_factory
+        >>> ep = ExecutionProfile(row_factory=tuple_factory)
+        >>> cluster = Cluster(execution_profiles={EXEC_PROFILE_DEFAULT: ep})
         >>> session = cluster.connect('mykeyspace')
-        >>> session.row_factory = tuple_factory
         >>> rows = session.execute("SELECT name, age FROM users LIMIT 1")
         >>> print rows[0]
         ('Bob', 42)
@@ -92,8 +93,9 @@ def named_tuple_factory(colnames, rows):
     Example::
 
         >>> from cassandra.query import named_tuple_factory
+        >>> ep = ExecutionProfile(row_factory=named_tuple_factory)
+        >>> cluster = Cluster(execution_profiles={EXEC_PROFILE_DEFAULT: ep})
         >>> session = cluster.connect('mykeyspace')
-        >>> session.row_factory = named_tuple_factory
         >>> rows = session.execute("SELECT name, age FROM users LIMIT 1")
         >>> user = rows[0]
 
@@ -122,7 +124,7 @@ def named_tuple_factory(colnames, rows):
                     "(see Python 'namedtuple' documentation for details on name rules). "
                     "Results will be returned with positional names. "
                     "Avoid this by choosing different names, using SELECT \"<col name>\" AS aliases, "
-                    "or specifying a different row_factory on your Session" %
+                    "or specifying a different row_factory on your ExecutionProfile" %
                     (colnames, clean_column_names))
         Row = namedtuple('Row', _sanitize_identifiers(clean_column_names))
 
@@ -136,8 +138,9 @@ def dict_factory(colnames, rows):
     Example::
 
         >>> from cassandra.query import dict_factory
+        >>> ep = ExecutionProfile(row_factory=dict_factory)
+        >>> cluster = Cluster(execution_profiles={EXEC_PROFILE_DEFAULT: ep})
         >>> session = cluster.connect('mykeyspace')
-        >>> session.row_factory = dict_factory
         >>> rows = session.execute("SELECT name, age FROM users LIMIT 1")
         >>> print rows[0]
         {u'age': 42, u'name': u'Bob'}
@@ -194,8 +197,7 @@ class Statement(object):
     keyspace = None
     """
     The string name of the keyspace this query acts on. This is used when
-    :class:`~.TokenAwarePolicy` is configured for
-    :attr:`.Cluster.load_balancing_policy`
+    :class:`~.TokenAwarePolicy` is configured in the profile load balancing policy.
 
     It is set implicitly on :class:`.BoundStatement`, and :class:`.BatchStatement`,
     but must be set explicitly on :class:`.SimpleStatement`.