+++ tests/integration/cqlengine/connections/test_connection.py
@@ -21,7 +21,7 @@ except ImportError:
 from cassandra.cqlengine.models import Model
 from cassandra.cqlengine import columns, connection
 from cassandra.cqlengine.management import sync_table
-from cassandra.cluster import Cluster, _clusters_for_shutdown
+from cassandra.cluster import Cluster, _clusters_for_shutdown, ExecutionProfile, EXEC_PROFILE_DEFAULT
 from cassandra.query import dict_factory
 
 from tests.integration import PROTOCOL_VERSION, execute_with_long_wait_retry, local
@@ -77,7 +77,8 @@ class SeveralConnectionsTest(BaseCassEngTestCase):
         cls.keyspace1 = 'ctest1'
         cls.keyspace2 = 'ctest2'
         super(SeveralConnectionsTest, cls).setUpClass()
-        cls.setup_cluster = Cluster(protocol_version=PROTOCOL_VERSION)
+        ep = ExecutionProfile(row_factory=dict_factory)
+        cls.setup_cluster = Cluster(protocol_version=PROTOCOL_VERSION, execution_profiles={EXEC_PROFILE_DEFAULT: ep})
         cls.setup_session = cls.setup_cluster.connect()
         ddl = "CREATE KEYSPACE {0} WITH replication = {{'class': 'SimpleStrategy', 'replication_factor': '{1}'}}".format(cls.keyspace1, 1)
         execute_with_long_wait_retry(cls.setup_session, ddl)
@@ -94,11 +95,10 @@ class SeveralConnectionsTest(BaseCassEngTestCase):
         models.DEFAULT_KEYSPACE
 
     def setUp(self):
-        self.c = Cluster(protocol_version=PROTOCOL_VERSION)
+        ep = ExecutionProfile(row_factory=dict_factory)
+        self.c = Cluster(protocol_version=PROTOCOL_VERSION, execution_profiles={EXEC_PROFILE_DEFAULT: ep})
         self.session1 = self.c.connect(keyspace=self.keyspace1)
-        self.session1.row_factory = dict_factory
         self.session2 = self.c.connect(keyspace=self.keyspace2)
-        self.session2.row_factory = dict_factory
 
     def tearDown(self):
         self.c.shutdown()