+++ tests/integration/standard/test_policies.py
@@ -17,7 +17,7 @@ try:
 except ImportError:
     import unittest  # noqa
 
-from cassandra.cluster import Cluster, ExecutionProfile, ResponseFuture
+from cassandra.cluster import Cluster, ExecutionProfile, EXEC_PROFILE_DEFAULT, ResponseFuture
 from cassandra.policies import HostFilterPolicy, RoundRobinPolicy,  SimpleConvictionPolicy, \
     WhiteListRoundRobinPolicy, ConstantSpeculativeExecutionPolicy
 from cassandra.hosts import Host
@@ -53,10 +53,11 @@ class HostFilterPolicyTests(unittest.TestCase):
         all_hosts = {Host("127.0.0.{}".format(i), SimpleConvictionPolicy) for i in (1, 2, 3)}
 
         predicate = lambda host: host.address == contact_point if external_event else True
-        cluster = Cluster((contact_point,), load_balancing_policy=HostFilterPolicy(RoundRobinPolicy(),
-                                                                                 predicate=predicate),
-                          protocol_version=PROTOCOL_VERSION, topology_event_refresh_window=0,
-                          status_event_refresh_window=0)
+        ep = ExecutionProfile(load_balancing_policy=HostFilterPolicy(
+            RoundRobinPolicy(), predicate=predicate))
+        cluster = Cluster((contact_point,), execution_profiles={EXEC_PROFILE_DEFAULT: ep},
+            protocol_version=PROTOCOL_VERSION, topology_event_refresh_window=0,
+            status_event_refresh_window=0)
         session = cluster.connect(wait_for_all_pools=True)
 
         queried_hosts = set()